Base simplifier test
